backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
requests==2.32.3
openai==1.46.0
azure-cognitiveservices-speech==1.37.0
python-dotenv==1.0.1
httpx==0.27.2

